Refactor event_log_storage into class module
